backend/pyproject.toml
[project]
name = "django-ters"
version = "0.1.0"
description = "Backend app for TERS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "click==8.1.7",
    "colorama==0.4.6",
    "cssbeautifier==1.15.1",
    "django>=5.2",
    "django-cors-headers==4.4.0",
    "django-debug-toolbar==4.4.6",
    "django-rest-framework==0.1.0",
    "djangorestframework==3.15.2",
    "djoser>=2.3.1",
    "editorconfig==0.12.4",
    "html-tag-names==0.1.2",
    "html-void-elements==0.1.0",
    "jsbeautifier==1.15.1",
    "json5==0.9.25",
    "pathspec==0.12.1",
    "pillow==10.4.0",
    "psycopg[binary]>=3.2.6",
    "python-dotenv>=1.1.0",
    "pyyaml==6.0.1",
    "regex==2024.11.6",
    "six==1.16.0",
    "sqlparse==0.5.1",
    "tomli==2.0.1",
    "tqdm==4.66.4",
    "typing-extensions==4.12.2",
]
